data: break up content.js, flash.js, sorting-rule.js
