+++ src/data/things/flash/index.js
@@ -1,0 +1,3 @@
+export * from './Flash.js';
+export * from './FlashAct.js';
+export * from './FlashSide.js';
