+++ src/data/things/index.js
@@ -1,8 +1,11 @@
 // Not actually the entry point for #things - that's init.js in this folder.
 
 export * from './album/index.js';
+export * from './content/index.js';
+export * from './flash/index.js';
 export * from './group/index.js';
 export * from './homepage-layout/index.js';
+export * from './sorting-rule/index.js';
 
 export * from './AdditionalFile.js';
 export * from './AdditionalName.js';
@@ -16,7 +19,3 @@ export * from './NewsEntry.js';
 export * from './StaticPage.js';
 export * from './Track.js';
 export * from './WikiInfo.js';
-
-export * from './content.js';
-export * from './flash.js';
-export * from './sorting-rule.js';
