+++ src/data/things/sorting-rule/index.js
@@ -1,0 +1,3 @@
+export * from './SortingRule.js';
+export * from './ThingSortingRule.js';
+export * from './DocumentSortingRule.js';
